+++ Magic/CHANGELOG.md
@@ -15,6 +15,7 @@
  - Add `damage_source: direct` option to Damage action, will bypass vanilla damage cooldowns (bypasses everything)
  - Allow using @attributes in custom wand lore
  - Add `command_source` parameter to Selector action, can be player, console (default) or opped_player
+ - Add allowed_enchantments, blocked_enchantments and slot parameters to CheckInventory, for checking on a specific inventory slot
 
 # 8.3.3